measurements/requirements.txt
pyserial
WiringPi
gpiozero
requests
python-dotenv
schedule
